A realistic analytics project about bank direct marketing
Modelling and Evaluation

In order to model and test the data, we divided it into training and test sets. 70 % of the data was used to train the model and 30 % was allocated for testing purposes. We included a confusion matrix in the same code chunk as the individual models to serve as an evaluation metric.
The following modelling techniques were used to develop our algorithms:
•	Support Vector Machine (SVM): SVM identifies an optimal separating hyperplane which maximizes the margin between different classes of the training data.
•	Random Forest: Random forest is an ensemble learning method for classification by constructing a multitude of decision trees at the stage of training the model and then outputting the class prediction of individual trees.
•	Decision tree (DT):  DT is a popular classification technique that can graphically organize information about possibility, consequences and end value.
•	Naïve Bayes: Naïve Bayes is a simple probabilistic classifier which is based on Bayes’ theorem and independence assumptions.
We also made an attempt to use logistic regression, but noted that the random forest served as a better model.

Conclusion

Taking all of the above into consideration, our final recommendation to Universal Credit bank is to implement an optimization strategy for their next marketing campaign by using our random forest model to determine which potentials customers to target.